requirements.txt
# FastAPI Core
fastapi==0.124.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.45
alembic==1.17.2
psycopg2-binary==2.9.11
asyncpg==0.31.0

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
pydantic[email]==2.12.5
pydantic-settings==2.12.0

# Email
emails==0.6
jinja2==3.1.6

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
httpx==0.28.1

# Utilities
python-dateutil==2.9.0.post0
